core, examples: minimal harness-rollout contract + runnable example for online RL with TRL
