+++ examples/trl_harness_rollout/run.py
@@ -19,14 +19,12 @@ import time
 from concurrent.futures import ThreadPoolExecutor
 from pathlib import Path
 
+
 sys.path.insert(0, str(Path(__file__).resolve().parent))
-sys.path.insert(
-    0, str(Path(__file__).resolve().parents[2] / "src")
-)  # make `openenv` importable
 from generate import FakeGenerate, VLLMGenerate  # noqa: E402
 from harness import ProxyAgentSessionFactory, TASKS  # noqa: E402
-from openenv.core.harness.interception import InterceptionServer  # noqa: E402
-from openenv.core.harness.rollout_worker import HarnessRolloutWorker  # noqa: E402
+from interception import InterceptionServer  # noqa: E402
+from rollout_worker import HarnessRolloutWorker  # noqa: E402
 
 
 def _wait_for_vllm(url: str, timeout_s: float = 60) -> None:
